requirements.txt
allure-python-commons~=2.14.1
selenium~=4.31.0
pytest~=8.3.5